+++ packages/heisenberg/src/Matter.ts
@@ -1,4 +1,0 @@
-import { SyncAsync } from '@jamashita/anden-type';
-import { Nihil } from './Nihil';
-
-export type Matter<T> = Exclude<T, SyncAsync<Nihil>>;