+++ packages/heisenberg/src/index.ts
@@ -2,7 +2,5 @@ export * from './Absent';
 export * from './Heisenberg';
 export * from './HeisenbergError';
 export * from './Lost';
-export * from './Matter';
-export * from './Nihil';
 export * from './Present';
 export * from './Uncertain';