+++ packages/plan/src/Plan.ts
@@ -2,6 +2,6 @@ import { DestroyPlan } from './Destroy/DestroyPlan';
 import { MapPlan } from './Map/MapPlan';
 import { RecoveryPlan } from './Recovery/RecoveryPlan';
 
-export interface Plan<M, R> extends MapPlan<M>, RecoveryPlan<R>, DestroyPlan {
+export interface Plan<out M, out R> extends MapPlan<M>, RecoveryPlan<R>, DestroyPlan {
   // NOOP
 }