+++ packages/plan/src/index.ts
@@ -1,5 +1,6 @@
 export * from './Destroy/index';
 export * from './Map/index';
 export * from './Recovery/index';
-export * from './PassThroughPlan';
+export * from './PassPlan';
 export * from './Plan';
+export * from './SpoilPlan';