+++ packages/unscharferelation/src/IUnscharferelation.ts
@@ -1,20 +1,28 @@
-import { Consumer, Kind, Peek, Serializable, Supplier, SyncAsync, UnaryFunction } from '@jamashita/anden-type';
-import { Heisenberg, Matter, Nihil } from '@jamashita/genitore-heisenberg';
+import { Consumer, Kind, Peek, Serializable, Supplier, UnaryFunction } from '@jamashita/anden-type';
+import { Heisenberg } from '@jamashita/genitore-heisenberg';
 
-export type UReturnType<T> = SyncAsync<IUnscharferelation<T> | Nihil | T>;
+export type UReturnType<T> =
+  IUnscharferelation<T>
+  | PromiseLike<IUnscharferelation<T>>
+  | PromiseLike<null | undefined | void>
+  | PromiseLike<T>
+  | T
+  | null
+  | undefined
+  | void;
 
-export interface IUnscharferelation<P> extends Serializable {
-  get(): Promise<Matter<P>>;
+export interface IUnscharferelation<out P> extends Serializable {
+  get(): Promise<Exclude<P, null | undefined | void>>;
 
   ifAbsent(consumer: Consumer<void>): this;
 
   ifLost(consumer: Consumer<unknown>): this;
 
-  ifPresent(consumer: Consumer<Matter<P>>): this;
+  ifPresent(consumer: Consumer<Exclude<P, null | undefined | void>>): this;
 
-  map<Q = P>(mapper: UnaryFunction<Matter<P>, UReturnType<Q>>): IUnscharferelation<Q>;
+  map<Q = P>(mapper: UnaryFunction<Exclude<P, null | undefined | void>, UReturnType<Q>>): IUnscharferelation<Q>;
 
-  pass(accepted: Consumer<Matter<P>>, declined: Consumer<void>, thrown: Consumer<unknown>): this;
+  pass(accepted: Consumer<Exclude<P, null | undefined | void>>, declined: Consumer<void>, thrown: Consumer<unknown>): this;
 
   peek(peek: Peek): this;