+++ packages/superposition/CHANGELOG.md
@@ -1,3 +1,12 @@
+# [@jamashita/genitore-superposition-v2.8.0](https://github.com/jamashita/genitore/compare/@jamashita/genitore-superposition-v2.7.0...@jamashita/genitore-superposition-v2.8.0) (2022-04-15)
+
+
+### Features
+
+* 🎸 generic ([cf05eba](https://github.com/jamashita/genitore/commit/cf05eba5e9cc5d8217b49e188eecd7924e3674da))
+* 🎸 no constructor ([07e1168](https://github.com/jamashita/genitore/commit/07e1168c638fea3e5ed6d7e477bd351a6101a23f))
+* 🎸 packages updated ([dfa727d](https://github.com/jamashita/genitore/commit/dfa727d8df515bfbafdb1248f131671f634247d5))
+
 # [@jamashita/genitore-superposition-v2.7.0](https://github.com/jamashita/genitore/compare/@jamashita/genitore-superposition-v2.6.0...@jamashita/genitore-superposition-v2.7.0) (2022-04-14)